Extract generic code out of labeling task
