+++ website/src/components/Loading/LoadingScreen.jsx
@@ -1,7 +1,7 @@
 import { Progress } from "@chakra-ui/react";
 import { useColorMode } from "@chakra-ui/react";
 
-export const LoadingScreen = ({ text }) => {
+export const LoadingScreen = ({ text = "Loading..." } = {}) => {
   const { colorMode } = useColorMode();
   const mainClasses = colorMode === "light" ? "bg-slate-300 text-gray-800" : "bg-slate-900 text-white";
